Add Resource Based Authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishesFromRestaurantById/DeleteDishesFromRestaurantByIdCommandHandler.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesFromRestaurantById;
@@ -10,7 +12,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishesFromRestaurantById
 public class DeleteDishesFromRestaurantByIdCommandHandler(
     ILogger<DeleteDishesFromRestaurantByIdCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesFromRestaurantByIdCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesFromRestaurantByIdCommand>
 
 
 {
@@ -20,6 +23,8 @@ public class DeleteDishesFromRestaurantByIdCommandHandler(
         var restaurant = await restaurantsRepository.GetRestaurantByIdAsync(request.restaurantId);
         if (restaurant == null) throw new NotFoundException(nameof(Restaurant),request.restaurantId.ToString());
 
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+            throw new ForbidException();
         var dishesToDelete = restaurant.Dishes;
 
        await dishesRepository.DeleteDishesAsync(dishesToDelete);
